refactor(AdminInput): file and directory structure
