+++ frontend/src/components/AdminInput/Mark/Toggle/_toggle.scss
@@ -1,5 +1,5 @@
-@use '../../../pages/Admin/colors' as color;
-@use '../common-options' as common;
+@use '../../../../pages/Admin/colors' as color;
+@use '../../common-options' as common;
 
 $button-height: 1.2rem;
 $button-width: 2.5rem;
