+++ frontend/src/components/AdminInput/Shared/TickBox/TickBoxGroup.tsx
@@ -1,7 +1,7 @@
 import { FC, HTMLAttributes } from 'react';
 
-const BoxGroup: FC<HTMLAttributes<HTMLElement>> = ({ children }) => (
+const TickBoxGroup: FC<HTMLAttributes<HTMLElement>> = ({ children }) => (
   <div className="admin-input__box-group">{children}</div>
 );
 
-export default BoxGroup;
+export default TickBoxGroup;
