+++ frontend/src/components/AdminInput/Shared/TickBox/_box.scss
@@ -1,5 +1,5 @@
-@use '../../../pages/Admin/colors' as color;
-@use '../common-options' as common;
+@use '../../../../pages/Admin/colors' as color;
+@use '../../common-options' as common;
 
 $checkbox-size: 0.9rem;
