+++ frontend/src/components/AdminInput/Shared/Title/_title.scss
@@ -1,4 +1,4 @@
-@use '../../../pages/Admin/colors' as color;
+@use '../../../../pages/Admin/colors' as color;
 
 .admin-input__title {
   font-size: 1.1rem;
