+++ frontend/src/components/AdminInput/common.ts
@@ -1,0 +1,10 @@
+export interface InputProps {
+  title: string;
+  required?: boolean;
+}
+
+export interface Placeholder {
+  placeholder: string;
+}
+
+export type CommonInputProps = InputProps & Placeholder;
